requirements.txt
aiogram==2.23.1
google-generativeai
python-dotenv
